+++ Structures/Buildings/Fisher.cs
@@ -25,7 +25,7 @@ namespace Shared.Structures
 
         public Fisher(
             HexCell Cell,
-            byte Tribe,
+            int Tribe,
             byte Level,
             byte Health,
             int TroopCount,
