+++ Structures/InventoryBuilding.cs
@@ -23,7 +23,7 @@ namespace Shared.Structures
 
         public InventoryBuilding(
             HexCell Cell, 
-            byte Tribe,
+            int Tribe,
             byte Level, 
             byte Health, 
             int TroopCount,
